Moved dask I/O code to dedicated submodule
